feat: created a utils module dedicated to historical forecast, addressed some review comments
