+++ darts/models/forecasting/forecasting_model.py
@@ -31,11 +31,10 @@ from darts import metrics
 from darts.dataprocessing.encoders import SequentialEncoder
 from darts.logging import get_logger, raise_if, raise_if_not, raise_log
 from darts.timeseries import TimeSeries
-from darts.utils import (
-    _build_tqdm_iterator,
+from darts.utils import _build_tqdm_iterator, _parallel_apply, _with_sanity_checks
+from darts.utils.historical_forecasts.utils import (
     _historical_forecasts_general_checks,
-    _parallel_apply,
-    _with_sanity_checks,
+    _historical_forecasts_start_warnings,
 )
 from darts.utils.timeseries_generation import (
     _build_forecast_series,
@@ -585,6 +584,11 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
             else:
                 intersect_ = tmp_
 
+        # end comes before the start
+        if intersect_[1] < intersect_[0]:
+            return None
+
+        # generate an index
         if not reduce_to_bounds:
             intersect_ = generate_index(
                 start=intersect_[0], end=intersect_[1], freq=series.freq
@@ -636,41 +640,6 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
         series = args[0]
         _historical_forecasts_general_checks(series, kwargs)
 
-    def _historical_forecasts_start_warnings(
-        self,
-        idx: int,
-        start: Union[pd.Timestamp, int],
-        start_time_: Union[int, pd.Timestamp],
-        historical_forecasts_time_index: Union[
-            pd.DatetimeIndex, pd.RangeIndex, Tuple[Any, Any]
-        ],
-    ):
-        """Warnings when start value provided by user is not within the forecastable indexes boundaries"""
-        if not isinstance(start, pd.Timestamp):
-            start_value_msg = (
-                f"value `{start}` corresponding to timestamp `{start_time_}`"
-            )
-        else:
-            start_value_msg = f"time `{start_time_}`"
-
-        if start_time_ < historical_forecasts_time_index[0]:
-            logger.warning(
-                f"`start` {start_value_msg} is before the first predictable/trainable historical "
-                f"forecasting point for series at index: {idx}. Ignoring `start` for this series and "
-                f"beginning at first trainable/predictable time: {historical_forecasts_time_index[0]}. "
-                f"To hide these warnings, set `show_warnings=False`."
-            )
-        else:
-            logger.warning(
-                f"`start` {start_value_msg} is after the last trainable/predictable historical "
-                f"forecasting point for series at index: {idx}. This would results in empty historical "
-                f"forecasts. Ignoring `start` for this series and beginning at first trainable/"
-                f"predictable time: {historical_forecasts_time_index[0]}. Non-empty forecasts can be "
-                f"generated by setting `start` value to times between (including): "
-                f"{historical_forecasts_time_index[0], historical_forecasts_time_index[-1]}. "
-                f"To hide these warnings, set `show_warnings=False`."
-            )
-
     def _get_last_prediction_time(self, series, forecast_horizon, overlap_end):
         if overlap_end:
             last_valid_pred_time = series.time_index[-1]
@@ -678,6 +647,15 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
             last_valid_pred_time = series.time_index[-forecast_horizon]
 
         return last_valid_pred_time
+
+    def _check_optimisable_historical_forecasts(
+        self,
+        forecast_horizon: int,
+        retrain: Union[bool, int, Callable[..., bool]],
+        show_warnings=bool,
+    ) -> bool:
+        """By default, historical forecasts cannot be optimized"""
+        return False
 
     @_with_sanity_checks("_historical_forecasts_sanity_checks")
     def historical_forecasts(
@@ -695,7 +673,7 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
         last_points_only: bool = True,
         verbose: bool = False,
         show_warnings: bool = True,
-        with_optimisation: bool = True,
+        enable_optimisation: bool = True,
     ) -> Union[
         TimeSeries, List[TimeSeries], Sequence[TimeSeries], Sequence[List[TimeSeries]]
     ]:
@@ -788,7 +766,10 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
         verbose
             Whether to print progress.
         show_warnings
-            Whether to show warnings related to parameters `start`, and `train_length`.
+            Whether to show warnings related to historical forecasts optimization, or parameters `start` and
+            `train_length`.
+        enable_optimisation
+            Whether to use the optimized version of historical_forecasts when supported and available.
 
         Returns
         -------
@@ -905,15 +886,15 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
         future_covariates = series2seq(future_covariates)
 
         if (
-            with_optimisation
-            and retrain is False
+            enable_optimisation
             and model.supports_optimized_historical_forecasts
-            and (
-                model.output_chunk_length is not None
-                and forecast_horizon <= model.output_chunk_length
+            and model._check_optimisable_historical_forecasts(
+                forecast_horizon=forecast_horizon,
+                retrain=retrain,
+                show_warnings=show_warnings,
             )
         ):
-            return model._optimized_historical_forecasts(
+            return model._optimised_historical_forecasts(
                 series=series,
                 past_covariates=past_covariates,
                 future_covariates=future_covariates,
@@ -1071,7 +1052,7 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
                     <= historical_forecasts_time_index[-1]
                 ):
                     if show_warnings:
-                        self._historical_forecasts_start_warnings(
+                        _historical_forecasts_start_warnings(
                             idx=idx,
                             start=start,
                             start_time_=start_time_,
@@ -2009,7 +1990,7 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
             if p.name != "self"
         }
 
-    def _optimized_historical_forecasts(
+    def _optimised_historical_forecasts(
         self,
         series: Sequence[TimeSeries],
         past_covariates: Optional[Sequence[TimeSeries]] = None,
@@ -2025,7 +2006,9 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
     ) -> Union[
         TimeSeries, List[TimeSeries], Sequence[TimeSeries], Sequence[List[TimeSeries]]
     ]:
-        pass
+        logger.warning(
+            "`optimised historical forecasts is not available for this model, use `historical_forecasts` instead."
+        )
 
 
 class LocalForecastingModel(ForecastingModel, ABC):
