+++ darts/utils/__init__.py
@@ -4,7 +4,6 @@ Utils
 """
 from .utils import (
     _build_tqdm_iterator,
-    _historical_forecasts_general_checks,
     _parallel_apply,
     _with_sanity_checks,
     retain_period_common_to_all,
