+++ darts/utils/historical_forecasts/__init__.py
@@ -1,0 +1,9 @@
+from .optimised_historical_forecasts import (
+    _optimised_historical_forecasts_regression_all_points,
+    _optimised_historical_forecasts_regression_last_points_only,
+)
+from .utils import (
+    _get_historical_forecast_boundaries,
+    _historical_forecasts_general_checks,
+    _historical_forecasts_start_warnings,
+)
